Add skipTar param of package.sh

+++ build/package.sh
@@ -1,21 +1,29 @@
 #!/usr/bin/env bash
 
 set -e
-
+is_skip_tar="${1:-}"
 # Notebook absolute path
 root_dir=$(cd -P -- "$(dirname -- "$0")/.." && pwd -P)
-cd ${root_dir}
+if [[ -n "${BYZER_NOTEBOOK_HOME}" ]]; then
+   root_dir=${BYZER_NOTEBOOK_HOME}
+fi
 
-echo ${root_dir}
+cd ${root_dir} && echo ${root_dir}
 
-# build front
-git clone https://github.com/byzer-org/byzer-notebook-vue.git
+if [[ ! -d byzer-notebook-vue/.git ]]; then
+    echo "cloning byzer-notebook-vue repo..."
+    # build front
+    git clone https://github.com/byzer-org/byzer-notebook-vue.git
+else
+    echo "update byzer-notebook-vue to latest..."
+    ( cd byzer-notebook-vue && git checkout main && git pull -r origin main )
+fi
 
 cd ${root_dir}/byzer-notebook-vue && bash ./build/build.sh
 
-console_static_resources_dir=${root_dir}/src/main/resources/static
+console_static_resources_dir=${root_dir}/src/main/resources/static &&
 echo ${console_static_resources_dir}
-[[ ! -d ${console_static_resources_dir} ]] && mkdir -p ${console_static_resources_dir} || rm -rf ${console_static_resources_dir}/*
+[[ ! -d ${console_static_resources_dir} ]] && mkdir -p ${console_static_resources_dir} || rm -rf "${console_static_resources_dir:?}"/*
 
 tar -xvf ${root_dir}/byzer-notebook-vue/build/*.tar.gz -C ${console_static_resources_dir}
 
@@ -53,24 +61,30 @@ echo `git rev-parse HEAD` | tee commit_SHA1
 echo "${version}" > VERSION
 
 
-## 5. copy console jar file
+## 4. copy console jar file
 cp ${root_dir}/target/notebook-console.jar .
 
-## 6. copy config
+## 5. copy config
 mkdir conf
 cp ${root_dir}/conf/notebook.properties.example conf/notebook.properties
 
-## 7. copy scripts
+## 6. copy scripts
 cp ${root_dir}/build/startup.sh .
 cp ${root_dir}/build/shutdown.sh .
 
-## 8. others
+## 7. others
 mkdir logs
 cp -r ${root_dir}/sample .
 
 cd ${root_dir}/dist
-tar -zcvf ${package_name}.tar.gz ${package_name}
 
-echo "====================================="
-echo "Build Finished!"
-echo "Location: ${root_dir}/dist/${package_name}.tar.gz"
+if [[ "${is_skip_tar}" == "skipTar" ]]; then
+  echo "====================================="
+  echo "Build Finished!"
+  echo "Location: ${root_dir}/dist/${package_name}"
+else
+  tar -zcvf ${package_name}.tar.gz ${package_name}
+  echo "====================================="
+  echo "Build Finished!"
+  echo "Location: ${root_dir}/dist/${package_name}.tar.gz"
+fi
